+++ R/rgho.R
@@ -27,11 +27,6 @@
 #'
 #' @importFrom curl ie_get_proxy_for_url
 #'
-#' @importFrom xml2 xml_find_all
-#' @importFrom xml2 xml_attr
-#' @importFrom xml2 xml_contents
-#' @importFrom xml2 xml_text
-#'
 #' @importFrom readr read_csv
 #'
 #' @importFrom magrittr "%>%"
@@ -39,16 +34,21 @@
 #' @importFrom magrittr "%<>%"
 #' @importFrom magrittr "extract2"
 #'
-#' @importFrom dplyr data_frame
+#' @importFrom tibble as_tibble
+#'
 #' @importFrom dplyr filter_
 #' @importFrom dplyr bind_rows
+#' @importFrom dplyr select_
 #'
 #' @importFrom tidyr spread_
 #'
 #' @importFrom lazyeval lazy_dots
 #'
 #' @importFrom stats runif
+#' @importFrom stats setNames
 #'
 #' @importFrom utils head
+#'
+#' @importFrom jsonlite fromJSON
 #'
 NULL